requirements.txt
spacy
attrs